requirements.txt
pillow~=10.1.0
torchvision~=0.17.0.dev20231106
numpy~=1.26.0
matplotlib~=3.8.1